bitvmx_transaction_monitor::monitor
Struct MockMonitorApi

pub struct MockMonitorApi { /* private fields */ }

Implementations

impl MockMonitorApi

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockMonitorApi

pub fn expect_is_ready(&mut self) -> &mut Expectation

Create an Expectation for mocking the is_ready method

pub fn expect_tick(&mut self) -> &mut Expectation

Create an Expectation for mocking the tick method

pub fn expect_get_monitor_height(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_monitor_height method

pub fn expect_get_current_block(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_current_block method

pub fn expect_get_confirmation_threshold(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_confirmation_threshold method

pub fn expect_monitor(&mut self) -> &mut Expectation

Create an Expectation for mocking the monitor method

pub fn expect_cancel(&mut self) -> &mut Expectation

Create an Expectation for mocking the cancel method

pub fn expect_get_news(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_news method

pub fn expect_ack_news(&mut self) -> &mut Expectation

Create an Expectation for mocking the ack_news method

pub fn expect_get_tx_status(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_tx_status method

pub fn expect_get_estimated_fee_rate(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_estimated_fee_rate method