bitvmx_transaction_monitor::store
Struct MockMonitorStore

pub struct MockMonitorStore { /* private fields */ }

Implementations

impl MockMonitorStore

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockMonitorStore

pub fn expect_get_monitor_height(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_monitor_height method

pub fn expect_update_monitor_height(&mut self) -> &mut Expectation

Create an Expectation for mocking the update_monitor_height method

pub fn expect_get_news(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_news method

pub fn expect_update_news(&mut self) -> &mut Expectation

Create an Expectation for mocking the update_news method

pub fn expect_ack_news(&mut self) -> &mut Expectation

Create an Expectation for mocking the ack_news method

pub fn expect_get_monitors(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_monitors method

pub fn expect_add_monitor(&mut self) -> &mut Expectation

Create an Expectation for mocking the add_monitor method

pub fn expect_deactivate_monitor(&mut self) -> &mut Expectation

Create an Expectation for mocking the deactivate_monitor method

pub fn expect_cancel_monitor(&mut self) -> &mut Expectation

Create an Expectation for mocking the cancel_monitor method